Working on integration. Still need to deal with null pointer exception during initialization of Tower  object. Also, there are multiple Tower classes, which should really be brought together, or their distinction should be clarified.

+++ core/src/com/week1/game/TowerBuilder/TowerBuilderStage.java
@@ -35,8 +35,12 @@ public class TowerBuilderStage {
 
     private void setWidgets() {
         sw = new StatsWidget();
-        sw.setLblTxt(builder.getCurrTower().getHp(), builder.getCurrTower().getAtk(),
-                builder.getCurrTower().getRange(), builder.getCurrTower().getPrice());
+        sw.setLblTxt(
+                (int)builder.getCurrTower().getHp(), 
+                (int)builder.getCurrTower().getAtk(),
+                (int)builder.getCurrTower().getRange(), 
+                (int)builder.getCurrTower().getPrice()
+        );
 
 
         for (int i = 0; i < builder.presets.NUM_PRESETS; i ++) {
@@ -79,7 +83,12 @@ public class TowerBuilderStage {
                 public void clicked(InputEvent event, float x, float y) {
                     Tower currTower = builder.presets.getTower(finalI + 1);
                     builder.setCurrTower(currTower);
-                    sw.setLblTxt(currTower.getHp(), currTower.getAtk(), currTower.getRange(), currTower.getPrice());
+                    sw.setLblTxt(
+                            (int)currTower.getHp(), 
+                            (int)currTower.getAtk(), 
+                            (int)currTower.getRange(), 
+                            (int)currTower.getPrice()
+                    );
                 }
             });
